refactor: KnowHowBox 컴포넌트화
